PIC_IRQ_STATUS 값을 받아오는 변수의 타입변경. 레지스터의 크기가 32비트임. 기존 unsigned long int -> 변경 후 unsigned int

+++ lesson_07/main.c
@@ -21,7 +21,7 @@ void swiHandler(unsigned int syscallnum)
 
 void irqHandler(void)
 {
-	unsigned long int pic_irq_status = *PIC_IRQ_STATUS;
+	unsigned int pic_irq_status = *PIC_IRQ_STATUS;
 
 	if (pic_irq_status)
 	{
